+++ 2024/src/Day_11.cs
@@ -8,14 +8,14 @@ public class Day_11 : IDay
         _logger = factory.CreateLogger<Day_11>();
     }
 
-    public int Solve(string inputPath)
+    public long Solve(string inputPath)
     {
         string[] lines = File.ReadAllLines(inputPath);
 
         return 0;
     }
 
-    public int SolvePartTwo(string inputPath)
+    public long SolvePartTwo(string inputPath)
     {
         string[] lines = File.ReadAllLines(inputPath);
