+++ 2024/test/unit/Day_11Test.cs
@@ -9,7 +9,7 @@ public class Day_11Test
         using ILoggerFactory factory = LoggerFactory.Create(builder => { });
         Day_11 solver = new(factory);
 
-        int result = solver.Solve("day_11_sample.txt");
+        long result = solver.Solve("day_11_sample.txt");
 
         Assert.Equal(11, result);
     }
@@ -20,7 +20,7 @@ public class Day_11Test
         using ILoggerFactory factory = LoggerFactory.Create(builder => { });
         Day_11 solver = new(factory);
 
-        int result = solver.SolvePartTwo("day_11_sample.txt");
+        long result = solver.SolvePartTwo("day_11_sample.txt");
 
         Assert.Equal(31, result);
     }
