+++ 2024/test/unit/Day_16Test.cs
@@ -9,7 +9,7 @@ public class Day_16Test
         using ILoggerFactory factory = LoggerFactory.Create(builder => { });
         Day_16 solver = new(factory);
 
-        int result = solver.Solve("day_16_sample.txt");
+        long result = solver.Solve("day_16_sample.txt");
 
         Assert.Equal(11, result);
     }
@@ -20,7 +20,7 @@ public class Day_16Test
         using ILoggerFactory factory = LoggerFactory.Create(builder => { });
         Day_16 solver = new(factory);
 
-        int result = solver.SolvePartTwo("day_16_sample.txt");
+        long result = solver.SolvePartTwo("day_16_sample.txt");
 
         Assert.Equal(31, result);
     }
