+++ 2024/test/unit/Day_3Test.cs
@@ -9,7 +9,7 @@ public class Day_3Test
         using ILoggerFactory factory = LoggerFactory.Create(builder => { });
         Day_3 solver = new(factory);
 
-        int result = solver.Solve("day_3_sample.txt");
+        long result = solver.Solve("day_3_sample.txt");
 
         Assert.Equal(161, result);
     }
@@ -20,7 +20,7 @@ public class Day_3Test
         using ILoggerFactory factory = LoggerFactory.Create(builder => { });
         Day_3 solver = new(factory);
 
-        int result = solver.SolvePartTwo("day_3p2_sample.txt");
+        long result = solver.SolvePartTwo("day_3p2_sample.txt");
 
         Assert.Equal(48, result);
     }
